Рефакторинг тестов
