+++ src/test/java/ru/yandex/practicum/filmorate/FilmorateApplicationTests.java
@@ -9,5 +9,4 @@ class FilmorateApplicationTests {
 	@Test
 	void contextLoads() {
 	}
-
 }
