+++ src/test/java/ru/yandex/practicum/filmorate/validator/FilmValidatorTest.java
@@ -1,18 +1,17 @@
-package ru.yandex.practicum.filmorate.controller;
+package ru.yandex.practicum.filmorate.validator;
 
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.validator.FilmValidator;
 
 import java.time.LocalDate;
 
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-public class FilmControllerTest {
+public class FilmValidatorTest {
 
     private Film.FilmBuilder builder;
     private Film film;
@@ -40,7 +39,7 @@ public class FilmControllerTest {
     public void shouldNotThrowExceptionWhenReleaseDateIs28_12_1895() {
         film = builder.name("Титаник").description("Все умерли")
                 .releaseDate(LocalDate.of(1895, 12, 28)).duration(194).build();
-        assertDoesNotThrow( () -> FilmValidator.validate(film));
+        assertDoesNotThrow(() -> FilmValidator.validate(film));
     }
 
     @Test
@@ -54,7 +53,7 @@ public class FilmControllerTest {
     public void shouldNotThrowExceptionWhenDurationIsPositive() {
         film = builder.name("Титаник").description("Все умерли")
                 .releaseDate(LocalDate.of(1997, 12, 19)).duration(1).build();
-        assertDoesNotThrow( () -> FilmValidator.validate(film));
+        assertDoesNotThrow(() -> FilmValidator.validate(film));
     }
 
     @Test
@@ -75,7 +74,7 @@ public class FilmControllerTest {
     public void shouldNotThrowExceptionWhenDescriptionIsNotLonger200Symbols() {
         film = builder.name("Титаник").description(StringUtils.repeat("a", 200))
                 .releaseDate(LocalDate.of(1997, 12, 19)).duration(1).build();
-        assertDoesNotThrow( () -> FilmValidator.validate(film));
+        assertDoesNotThrow(() -> FilmValidator.validate(film));
     }
 
     @Test
